+++ src/utils.rs
@@ -35,6 +35,7 @@ pub mod on_drop_event;
 pub mod once;
 pub mod opaque;
 pub mod opaque_cell;
+pub mod opt;
 pub mod option_ext;
 pub mod oserror;
 pub mod page_size;