+++ release-notes.md
@@ -5,6 +5,7 @@
 - Implement wl-fixes.
 - Implement ei_touchscreen v2.
 - Implement idle-notification v2.
+- Implement ext-workspaces.
 
 # 1.7.0 (2024-10-25)